Adds favories_create facade and refactors endpoint to use facade
